pyproject.toml
[project]
name = "love-guard-api"
version = "2.0.0"
description = "Romance Scam Prevention API with DDD Architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",

    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # HTTP Client
    "httpx>=0.28.0",

    # Image Processing
    "pillow>=11.0.0",
    "numpy>=2.0.0",

    # AI/ML
    "deepface>=0.0.93",
    "tensorflow>=2.18.0",
    "openai>=1.59.0",

    # Database
    "qdrant-client>=1.12.0",
    "neo4j>=5.27.0",

    # Web Scraping
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    # Linting & Formatting
    "ruff>=0.9.0",

    # Type Checking
    "basedpyright>=1.23.0",

    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",

    # Pre-commit hooks
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ==================== Ruff 설정 ====================
[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
    ".ruff_cache",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PERF",   # perflint (performance)
    "RUF",    # ruff-specific rules
    "ASYNC",  # async-related rules
    "S",      # bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
    "S101",   # assert used (OK in tests)
    "S104",   # possible binding to all interfaces
    "S105",   # hardcoded password (false positives)
    "S106",   # hardcoded password (false positives)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ==================== Basedpyright 설정 ====================
[tool.basedpyright]
pythonVersion = "3.11"
pythonPlatform = "All"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportUnusedFunction = false
reportOptionalMemberAccess = false
reportGeneralTypeIssues = true
reportPrivateUsage = false
venvPath = "."
venv = ".venv"
include = ["src"]
exclude = [
    ".venv",
    "**/__pycache__",
]

# ==================== Pytest 설정 ====================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
